# backend/openapi_generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AGRO",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/rejection-entries/": {
      "post": {
        "tags": [
          "Rejection Entries"
        ],
        "summary": "Create Route",
        "operationId": "create_route_v1_rejection_entries__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectionEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectionEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Rejection Entries"
        ],
        "summary": "List rejections",
        "description": "Retrieve all rejection entries.\n\nArgs:\n    db (Session): Database session dependency.\n\nReturns:\n    List[RejectionEntryRead]: List of rejections.",
        "operationId": "read_all_v1_rejection_entries__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RejectionEntryRead"
                  },
                  "title": "Response Read All V1 Rejection Entries  Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rejection-entries/list": {
      "get": {
        "tags": [
          "Rejection Entries"
        ],
        "summary": "Filter rejections",
        "description": "Retrieve rejection entries filtered by date and item IDs.\n\nArgs:\n    rejection_date (date): Date to filter rejections.\n    item_ids (Optional[List[int]]): List of item IDs to filter.\n    db (Session): Database session dependency.\n\nReturns:\n    List[RejectionEntryRead]: List of filtered rejections.",
        "operationId": "get_filtered_rejections_v1_rejection_entries_list_get",
        "parameters": [
          {
            "name": "rejection_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Rejection date",
              "title": "Rejection Date"
            },
            "description": "Rejection date"
          },
          {
            "name": "item_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by item IDs",
              "title": "Item Ids"
            },
            "description": "Filter by item IDs"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RejectionEntryRead"
                  },
                  "title": "Response Get Filtered Rejections V1 Rejection Entries List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/rejection-entries/rejection-entries/": {
      "post": {
        "tags": [
          "Rejection Entry",
          "Rejection Entries"
        ],
        "summary": "Create Route",
        "operationId": "create_route_v1_rejection_entries_rejection_entries__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectionEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectionEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Rejection Entry",
          "Rejection Entries"
        ],
        "summary": "List rejections",
        "description": "Retrieve all rejection entries.\n\nArgs:\n    db (Session): Database session dependency.\n\nReturns:\n    List[RejectionEntryRead]: List of rejections.",
        "operationId": "read_all_v1_rejection_entries_rejection_entries__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RejectionEntryRead"
                  },
                  "title": "Response Read All V1 Rejection Entries Rejection Entries  Get"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchRead": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "expiry_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "item_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Name"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "quantity",
          "unit",
          "id"
        ],
        "title": "BatchRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RejectionEntryCreate": {
        "properties": {
          "batch_id": {
            "type": "integer",
            "title": "Batch Id"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "rejection_date": {
            "type": "string",
            "format": "date",
            "title": "Rejection Date"
          },
          "rejected_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejected By"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "quantity",
          "reason",
          "rejection_date",
          "rejected_by"
        ],
        "title": "RejectionEntryCreate"
      },
      "RejectionEntryRead": {
        "properties": {
          "batch_id": {
            "type": "integer",
            "title": "Batch Id"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "rejection_date": {
            "type": "string",
            "format": "date",
            "title": "Rejection Date"
          },
          "rejected_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejected By"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "ISO 8601 format"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "ISO 8601 format"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "batch": {
            "$ref": "#/components/schemas/BatchRead"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "quantity",
          "reason",
          "rejection_date",
          "rejected_by",
          "id",
          "unit",
          "created_at",
          "updated_at",
          "created_by",
          "updated_by",
          "batch"
        ],
        "title": "RejectionEntryRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}